Add *data to oni_create_frame
- Previously, oni_create_frame relied on an secondary memcpy to fill its
data field with the date block to be written to a device. Considering
one_create_frame required a data_sz arguement, this seemed very strange.
This commit just addes the data/data_sz together when creating a frame.
In this way oni_create_frame and oni_read_frame are dual: one creates
frames from hardware and the other from user design and data.
- This is an API breaking change, so it requires a major version update.