Made a simple hardware emulation driver
- onidriver_test is a simple, ONI-compliant hardware emulator that can
be used for testing API implementations without physical hardware.
- Its very simple and has some limitations that might be worth improving
in the future:

1. It is single threaded (i.e. no background process generating data).
Therefore, "hardware" is lazy and API calls also perform data
generation. This means there are never issues with buffering
and finite, random register access times, which are real-world issues we
may want to model.
2. (Related to 1) The ONI_OPT_RUNNING context option does nothing. To
make it work, data would need to be produced on a separate thread and
passed it through a blocking FIFO
3. It only supports a device table with containing devices with a single
fixed read/write size. This is easy to fix if I want to put in the
effort.
4. The block read size must be a multiple of the frame size
5. Writing to devices is fine, but does nothing (data is just ignored)